issue #3: refactor routing
